Fix: Troca do tipo dos IDs de int para Long

+++ src/main/java/br/edu/fatecgru/mercado_inteligente/controller/FornecedorController.java
@@ -29,7 +29,7 @@ public class FornecedorController {
 
 	// Busca por ID
 	@GetMapping("/{id}")
-	public Fornecedor buscarPorId(@PathVariable int id) {
+	public Fornecedor buscarPorId(@PathVariable Long id) {
 		return fornecedorService.getById(id);
 	}
