+++ src/main/java/br/edu/fatecgru/mercado_inteligente/repository/CarrinhoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.fatecgru.mercado_inteligente.model.entity.Carrinho;
 
-public interface CarrinhoRepository extends JpaRepository<Carrinho, Integer> {
+public interface CarrinhoRepository extends JpaRepository<Carrinho, Long> {
 
 }
